Refactor ExpectedCheckHolder from interfaces to impl package
