+++ lib/src/main/java/com/auth0/jwt/impl/ExpectedCheckHolder.java
@@ -1,4 +1,7 @@
-package com.auth0.jwt.interfaces;
+package com.auth0.jwt.impl;
+
+import com.auth0.jwt.interfaces.Claim;
+import com.auth0.jwt.interfaces.DecodedJWT;
 
 /**
  * This holds the checks that are run to verify a JWT.
